requirements.txt
altgraph==0.16.1
anytree==2.6.0
boto3==1.9.183
botocore==1.12.183
configparser==3.5.0
macholib==1.11
pathtools==0.1.2
pytz==2018.5
urllib3==1.22
validators==0.12.0
xlrd==1.1.0
XlsxWriter==1.0.2
cassandra-driver==3.19.0
prometheus_client==0.7.1
psycopg2-binary==2.8.3
redis==3.3.11
gelf-formatter
psutil
confluent_kafka
